+++ packages/components/organisms/f-checkout/src/components/_tests/Checkout.test.js
@@ -8,7 +8,7 @@ import VueCheckout from '../Checkout.vue';
 import EventNames from '../../event-names';
 
 import {
-    defaultState, defaultActions, i18n, createStore, $logger
+    defaultCheckoutState, defaultCheckoutActions, i18n, createStore, $logger
 } from './helpers/setup';
 
 const localVue = createLocalVue();
@@ -149,7 +149,7 @@ describe('Checkout', () => {
             it('should display the address block if set to `delivery`', async () => {
                 // Act
                 const wrapper = shallowMount(VueCheckout, {
-                    store: createStore({ ...defaultState, serviceType: CHECKOUT_METHOD_DELIVERY }),
+                    store: createStore({ ...defaultCheckoutState, serviceType: CHECKOUT_METHOD_DELIVERY }),
                     i18n,
                     localVue,
                     propsData
@@ -164,7 +164,7 @@ describe('Checkout', () => {
             it('should not display the address block if set to `collection`', async () => {
                 // Act
                 const wrapper = shallowMount(VueCheckout, {
-                    store: createStore({ ...defaultState, serviceType: CHECKOUT_METHOD_COLLECTION }),
+                    store: createStore({ ...defaultCheckoutState, serviceType: CHECKOUT_METHOD_COLLECTION }),
                     i18n,
                     localVue,
                     propsData
@@ -261,7 +261,7 @@ describe('Checkout', () => {
 
                 // Act
                 shallowMount(VueCheckout, {
-                    store: createStore(defaultState, { ...defaultActions, setAuthToken }),
+                    store: createStore(defaultCheckoutState, { ...defaultCheckoutActions, setAuthToken }),
                     i18n,
                     localVue,
                     propsData
@@ -318,7 +318,7 @@ describe('Checkout', () => {
             it('should return `true` if `serviceType` is set to Delivery', () => {
                 // Arrange and Act
                 const wrapper = shallowMount(VueCheckout, {
-                    store: createStore({ ...defaultState, serviceType: CHECKOUT_METHOD_DELIVERY }),
+                    store: createStore({ ...defaultCheckoutState, serviceType: CHECKOUT_METHOD_DELIVERY }),
                     i18n,
                     localVue,
                     propsData
@@ -331,7 +331,7 @@ describe('Checkout', () => {
             it('should return `false` if `serviceType` is set to Collection', () => {
                 // Arrange and Act
                 const wrapper = shallowMount(VueCheckout, {
-                    store: createStore({ ...defaultState, serviceType: CHECKOUT_METHOD_COLLECTION }),
+                    store: createStore({ ...defaultCheckoutState, serviceType: CHECKOUT_METHOD_COLLECTION }),
                     i18n,
                     localVue,
                     propsData
@@ -374,7 +374,7 @@ describe('Checkout', () => {
                 // Arrange
                 const wrapper = shallowMount(VueCheckout, {
                     store: createStore({
-                        ...defaultState,
+                        ...defaultCheckoutState,
                         isLoggedIn: true,
                         serviceType: CHECKOUT_METHOD_DELIVERY,
                         address: {}
@@ -395,7 +395,7 @@ describe('Checkout', () => {
                 // Arrange
                 const wrapper = shallowMount(VueCheckout, {
                     store: createStore({
-                        ...defaultState,
+                        ...defaultCheckoutState,
                         isLoggedIn: false,
                         serviceType: CHECKOUT_METHOD_DELIVERY,
                         address: {}
@@ -416,7 +416,7 @@ describe('Checkout', () => {
                 // Arrange
                 const wrapper = shallowMount(VueCheckout, {
                     store: createStore({
-                        ...defaultState,
+                        ...defaultCheckoutState,
                         isLoggedIn: true,
                         serviceType: CHECKOUT_METHOD_COLLECTION,
                         address: {}
@@ -435,7 +435,7 @@ describe('Checkout', () => {
                 // Arrange
                 const wrapper = shallowMount(VueCheckout, {
                     store: createStore({
-                        ...defaultState,
+                        ...defaultCheckoutState,
                         isLoggedIn: true,
                         serviceType: CHECKOUT_METHOD_DELIVERY,
                         address: { line1: 'Fleet Place House', postcode: 'EC4M 7RF', city: 'London' }
@@ -455,14 +455,19 @@ describe('Checkout', () => {
     });
 
     describe('mounted ::', () => {
+        let initialiseSpy;
+        let trackInitialLoadSpy;
+
+        beforeEach(() => {
+            initialiseSpy = jest.spyOn(VueCheckout.methods, 'initialise');
+            trackInitialLoadSpy = jest.spyOn(VueCheckout.methods, 'trackInitialLoad');
+        });
+
         afterEach(() => {
             jest.clearAllMocks();
         });
 
         it('should call `initialise`', () => {
-            // Arrange
-            const initialiseSpy = jest.spyOn(VueCheckout.methods, 'initialise');
-
             // Act
             shallowMount(VueCheckout, {
                 store: createStore(),
@@ -473,6 +478,21 @@ describe('Checkout', () => {
 
             // Assert
             expect(initialiseSpy).toHaveBeenCalled();
+        });
+
+        it('should call `trackInitialLoad`', async () => {
+            // Act
+            const wrapper = shallowMount(VueCheckout, {
+                store: createStore(),
+                i18n,
+                localVue,
+                propsData
+            });
+
+            await wrapper.vm.initialise();
+
+            // Assert
+            expect(trackInitialLoadSpy).toHaveBeenCalled();
         });
     });
 
@@ -630,7 +650,7 @@ describe('Checkout', () => {
                     const loadCheckoutSpy = jest.spyOn(VueCheckout.methods, 'loadCheckout');
 
                     shallowMount(VueCheckout, {
-                        store: createStore({ ...defaultState, isLoggedIn: true }),
+                        store: createStore({ ...defaultCheckoutState, isLoggedIn: true }),
                         i18n,
                         localVue,
                         propsData
@@ -750,7 +770,7 @@ describe('Checkout', () => {
                 describe('when all the fields are populated correctly', () => {
                     beforeEach(() => {
                         wrapper = mount(VueCheckout, {
-                            store: createStore({ ...defaultState, serviceType: CHECKOUT_METHOD_COLLECTION }),
+                            store: createStore({ ...defaultCheckoutState, serviceType: CHECKOUT_METHOD_COLLECTION }),
                             i18n,
                             localVue,
                             propsData,
@@ -795,10 +815,10 @@ describe('Checkout', () => {
                     // Arrange
                     wrapper = mount(VueCheckout, {
                         store: createStore({
-                            ...defaultState,
+                            ...defaultCheckoutState,
                             serviceType: CHECKOUT_METHOD_COLLECTION,
                             customer: {
-                                ...defaultState.customer,
+                                ...defaultCheckoutState.customer,
                                 mobileNumber: ''
                             }
                         }),
@@ -824,10 +844,10 @@ describe('Checkout', () => {
                     // Arrange
                     wrapper = mount(VueCheckout, {
                         store: createStore({
-                            ...defaultState,
+                            ...defaultCheckoutState,
                             serviceType: CHECKOUT_METHOD_COLLECTION,
                             customer: {
-                                ...defaultState.customer,
+                                ...defaultCheckoutState.customer,
                                 mobileNumber: '077777'
                             }
                         }),
@@ -854,10 +874,10 @@ describe('Checkout', () => {
                     // Arrange
                     wrapper = mount(VueCheckout, {
                         store: createStore({
-                            ...defaultState,
+                            ...defaultCheckoutState,
                             serviceType: CHECKOUT_METHOD_COLLECTION,
                             customer: {
-                                ...defaultState.customer,
+                                ...defaultCheckoutState.customer,
                                 mobileNumber: 'hs;-j`$e&1l'
                             }
                         }),
@@ -883,7 +903,7 @@ describe('Checkout', () => {
                 it('should not create validations for address', () => {
                     // Arrange
                     wrapper = mount(VueCheckout, {
-                        store: createStore({ ...defaultState, serviceType: CHECKOUT_METHOD_COLLECTION }),
+                        store: createStore({ ...defaultCheckoutState, serviceType: CHECKOUT_METHOD_COLLECTION }),
                         i18n,
                         localVue,
                         propsData,
@@ -900,10 +920,10 @@ describe('Checkout', () => {
                     // Arrange
                     wrapper = mount(VueCheckout, {
                         store: createStore({
-                            ...defaultState,
+                            ...defaultCheckoutState,
                             serviceType: CHECKOUT_METHOD_DELIVERY,
                             customer: {
-                                ...defaultState.customer,
+                                ...defaultCheckoutState.customer,
                                 mobileNumber: ''
                             }
                         }),
@@ -929,7 +949,7 @@ describe('Checkout', () => {
                 describe('when all the fields are populated correctly', () => {
                     beforeEach(() => {
                         wrapper = mount(VueCheckout, {
-                            store: createStore({ ...defaultState, serviceType: CHECKOUT_METHOD_DELIVERY }),
+                            store: createStore({ ...defaultCheckoutState, serviceType: CHECKOUT_METHOD_DELIVERY }),
                             i18n,
                             localVue,
                             propsData,
@@ -974,10 +994,10 @@ describe('Checkout', () => {
                     // Arrange
                     wrapper = mount(VueCheckout, {
                         store: createStore({
-                            ...defaultState,
+                            ...defaultCheckoutState,
                             serviceType: CHECKOUT_METHOD_DELIVERY,
                             address: {
-                                ...defaultState.address,
+                                ...defaultCheckoutState.address,
                                 line1: ''
                             }
                         }),
@@ -1003,10 +1023,10 @@ describe('Checkout', () => {
                     // Arrange
                     wrapper = mount(VueCheckout, {
                         store: createStore({
-                            ...defaultState,
+                            ...defaultCheckoutState,
                             serviceType: CHECKOUT_METHOD_DELIVERY,
                             address: {
-                                ...defaultState.address,
+                                ...defaultCheckoutState.address,
                                 city: ''
                             }
                         }),
@@ -1032,10 +1052,10 @@ describe('Checkout', () => {
                     // Arrange
                     wrapper = mount(VueCheckout, {
                         store: createStore({
-                            ...defaultState,
+                            ...defaultCheckoutState,
                             serviceType: CHECKOUT_METHOD_DELIVERY,
                             address: {
-                                ...defaultState.address,
+                                ...defaultCheckoutState.address,
                                 postcode: ''
                             }
                         }),
@@ -1061,10 +1081,10 @@ describe('Checkout', () => {
                     // Arrange
                     wrapper = mount(VueCheckout, {
                         store: createStore({
-                            ...defaultState,
+                            ...defaultCheckoutState,
                             serviceType: CHECKOUT_METHOD_DELIVERY,
                             address: {
-                                ...defaultState.address,
+                                ...defaultCheckoutState.address,
                                 postcode: '?!hdb-se'
                             }
                         }),
@@ -1090,10 +1110,10 @@ describe('Checkout', () => {
                     // Arrange
                     wrapper = mount(VueCheckout, {
                         store: createStore({
-                            ...defaultState,
+                            ...defaultCheckoutState,
                             serviceType: CHECKOUT_METHOD_DELIVERY,
                             address: {
-                                ...defaultState.address,
+                                ...defaultCheckoutState.address,
                                 postcode: 'EC4M 7R'
                             }
                         }),
@@ -1126,10 +1146,10 @@ describe('Checkout', () => {
                     // Arrange
                     wrapper = mount(VueCheckout, {
                         store: createStore({
-                            ...defaultState,
+                            ...defaultCheckoutState,
                             serviceType: CHECKOUT_METHOD_DELIVERY,
                             address: {
-                                ...defaultState.address,
+                                ...defaultCheckoutState.address,
                                 postcode: 'EC4M 7R'
                             }
                         }),
@@ -1159,7 +1179,7 @@ describe('Checkout', () => {
                     const setupGuestUserSpy = jest.spyOn(VueCheckout.methods, 'setupGuestUser');
                     const wrapper = shallowMount(VueCheckout, {
                         store: createStore({
-                            ...defaultState,
+                            ...defaultCheckoutState,
                             isLoggedIn: false
                         }),
                         i18n,
@@ -1181,7 +1201,7 @@ describe('Checkout', () => {
                     // Arrange
                     const wrapper = mount(VueCheckout, {
                         store: createStore({
-                            ...defaultState,
+                            ...defaultCheckoutState,
                             serviceType: CHECKOUT_METHOD_COLLECTION,
                             isLoggedIn: false
                         }),
@@ -1205,10 +1225,10 @@ describe('Checkout', () => {
                     // Arrange
                     const wrapper = mount(VueCheckout, {
                         store: createStore({
-                            ...defaultState,
+                            ...defaultCheckoutState,
                             serviceType: CHECKOUT_METHOD_COLLECTION,
                             customer: {
-                                ...defaultState.customer,
+                                ...defaultCheckoutState.customer,
                                 firstName: ''
                             },
                             isLoggedIn: false
@@ -1235,10 +1255,10 @@ describe('Checkout', () => {
                     // Arrange
                     const wrapper = mount(VueCheckout, {
                         store: createStore({
-                            ...defaultState,
+                            ...defaultCheckoutState,
                             serviceType: CHECKOUT_METHOD_COLLECTION,
                             customer: {
-                                ...defaultState.customer,
+                                ...defaultCheckoutState.customer,
                                 lastName: ''
                             },
                             isLoggedIn: false
@@ -1265,10 +1285,10 @@ describe('Checkout', () => {
                     // Arrange
                     const wrapper = mount(VueCheckout, {
                         store: createStore({
-                            ...defaultState,
+                            ...defaultCheckoutState,
                             serviceType: CHECKOUT_METHOD_COLLECTION,
                             customer: {
-                                ...defaultState.customer,
+                                ...defaultCheckoutState.customer,
                                 email: ''
                             },
                             isLoggedIn: false
@@ -1295,10 +1315,10 @@ describe('Checkout', () => {
                     // Arrange
                     const wrapper = mount(VueCheckout, {
                         store: createStore({
-                            ...defaultState,
+                            ...defaultCheckoutState,
                             serviceType: CHECKOUT_METHOD_COLLECTION,
                             customer: {
-                                ...defaultState.customer,
+                                ...defaultCheckoutState.customer,
                                 email: '£Gs7asd263('
                             },
                             isLoggedIn: false
@@ -1328,7 +1348,7 @@ describe('Checkout', () => {
                     const setupGuestUserSpy = jest.spyOn(VueCheckout.methods, 'setupGuestUser');
                     const wrapper = shallowMount(VueCheckout, {
                         store: createStore({
-                            ...defaultState,
+                            ...defaultCheckoutState,
                             authToken: 'sampleToken',
                             isLoggedIn: true
                         }),
@@ -1351,7 +1371,7 @@ describe('Checkout', () => {
                     // Arrange
                     const wrapper = shallowMount(VueCheckout, {
                         store: createStore({
-                            ...defaultState,
+                            ...defaultCheckoutState,
                             authToken: 'sampleToken',
                             isLoggedIn: true
                         }),
@@ -1434,7 +1454,7 @@ describe('Checkout', () => {
                 const createGuestUserSpy = jest.spyOn(VueCheckout.methods, 'createGuestUser');
                 const wrapper = shallowMount(VueCheckout, {
                     store: createStore({
-                        ...defaultState,
+                        ...defaultCheckoutState,
                         customer
                     }),
                     i18n,
@@ -1463,7 +1483,7 @@ describe('Checkout', () => {
                     jest.spyOn(VueCheckout.methods, 'initialise').mockImplementation();
 
                     wrapper = mount(VueCheckout, {
-                        store: createStore(defaultState, { ...defaultActions, createGuestUser: jest.fn(async () => Promise.reject()) }),
+                        store: createStore(defaultCheckoutState, { ...defaultCheckoutActions, createGuestUser: jest.fn(async () => Promise.reject()) }),
                         i18n,
                         localVue,
                         propsData,
@@ -1532,7 +1552,7 @@ describe('Checkout', () => {
                     jest.spyOn(VueCheckout.methods, 'initialise').mockImplementation();
 
                     wrapper = mount(VueCheckout, {
-                        store: createStore(defaultState, { ...defaultActions, getCheckout: jest.fn(async () => Promise.reject()) }),
+                        store: createStore(defaultCheckoutState, { ...defaultCheckoutActions, getCheckout: jest.fn(async () => Promise.reject()) }),
                         i18n,
                         localVue,
                         propsData,
@@ -1598,7 +1618,7 @@ describe('Checkout', () => {
                     jest.spyOn(VueCheckout.methods, 'initialise').mockImplementation();
 
                     wrapper = mount(VueCheckout, {
-                        store: createStore(defaultState, { ...defaultActions, getAvailableFulfilment: jest.fn(async () => Promise.reject()) }),
+                        store: createStore(defaultCheckoutState, { ...defaultCheckoutActions, getAvailableFulfilment: jest.fn(async () => Promise.reject()) }),
                         i18n,
                         localVue,
                         propsData,
@@ -1705,7 +1725,7 @@ describe('Checkout', () => {
                     jest.spyOn(VueCheckout.methods, 'initialise').mockImplementation();
 
                     wrapper = mount(VueCheckout, {
-                        store: createStore(defaultState, { ...defaultActions, getGeoLocation: jest.fn(async () => Promise.reject()) }),
+                        store: createStore(defaultCheckoutState, { ...defaultCheckoutActions, getGeoLocation: jest.fn(async () => Promise.reject()) }),
                         i18n,
                         localVue,
                         propsData,
@@ -1739,7 +1759,7 @@ describe('Checkout', () => {
                     getGeoLocationSpy = jest.spyOn(VueCheckout.methods, 'getGeoLocation');
 
                     wrapper = mount(VueCheckout, {
-                        store: createStore({ ...defaultState, address: {} }),
+                        store: createStore({ ...defaultCheckoutState, address: {} }),
                         i18n,
                         localVue,
                         propsData,
@@ -1781,7 +1801,7 @@ describe('Checkout', () => {
                     jest.spyOn(VueCheckout.methods, 'initialise').mockImplementation();
 
                     wrapper = mount(VueCheckout, {
-                        store: createStore(defaultState, { ...defaultActions, getBasket: jest.fn(async () => Promise.reject()) }),
+                        store: createStore(defaultCheckoutState, { ...defaultCheckoutActions, getBasket: jest.fn(async () => Promise.reject()) }),
                         i18n,
                         localVue,
                         propsData,
@@ -1844,7 +1864,7 @@ describe('Checkout', () => {
                 it('should emit failure event and set `hasCheckoutLoadedSuccessfully` to `false`', async () => {
                     // Arrange
                     const wrapper = mount(VueCheckout, {
-                        store: createStore(defaultState, { ...defaultActions, getAddress: jest.fn(async () => Promise.reject()) }),
+                        store: createStore(defaultCheckoutState, { ...defaultCheckoutActions, getAddress: jest.fn(async () => Promise.reject()) }),
                         i18n,
                         localVue,
                         propsData
@@ -2040,18 +2060,48 @@ describe('Checkout', () => {
         describe('onFormSubmit ::', () => {
             let isFormValidSpy;
             let submitCheckoutSpy;
+            let trackFormInteractionSpy;
 
             beforeEach(() => {
                 isFormValidSpy = jest.spyOn(VueCheckout.methods, 'isFormValid');
                 submitCheckoutSpy = jest.spyOn(VueCheckout.methods, 'submitCheckout');
+                trackFormInteractionSpy = jest.spyOn(VueCheckout.methods, 'trackFormInteraction');
             });
 
             afterEach(() => {
                 jest.clearAllMocks();
             });
 
-            it('should emit `CheckoutValidationError` with validation state if form is invalid', async () => {
+            it('should call `trackFormInteractionSpy` with correct `action`', async () => {
                 // Arrange
+                isFormValidSpy.mockReturnValue(true);
+
+                const wrapper = mount(VueCheckout, {
+                    store: createStore(),
+                    i18n,
+                    localVue,
+                    propsData,
+                    mocks: {
+                        $v,
+                        $logger
+                    }
+                });
+
+                const payload = {
+                    action: 'submit'
+                };
+
+                // Act
+                await wrapper.vm.onFormSubmit();
+
+                // Assert
+                expect(trackFormInteractionSpy).toHaveBeenCalledWith(payload);
+            });
+
+            describe('when the form is Invalid', () => {
+                let wrapper;
+                let getFormValidationStateSpy;
+
                 const mockValidationState = {
                     validFields: [
                         'customer.mobileNumber',
@@ -2062,76 +2112,117 @@ describe('Checkout', () => {
                     invalidFields: []
                 };
 
-                const getFormValidationStateSpy = jest.spyOn(validations, 'getFormValidationState');
+                beforeEach(() => {
+                    isFormValidSpy.mockReturnValue(false);
 
-                getFormValidationStateSpy.mockReturnValue(mockValidationState);
-                isFormValidSpy.mockReturnValue(false);
+                    getFormValidationStateSpy = jest.spyOn(validations, 'getFormValidationState');
+                    getFormValidationStateSpy.mockReturnValue(mockValidationState);
 
-                const wrapper = mount(VueCheckout, {
-                    store: createStore(),
-                    i18n,
-                    localVue,
-                    propsData,
-                    mocks: {
-                        $v,
-                        $logger
-                    }
+                    wrapper = mount(VueCheckout, {
+                        store: createStore(),
+                        i18n,
+                        localVue,
+                        propsData,
+                        mocks: {
+                            $v,
+                            $logger
+                        }
+                    });
                 });
 
-                // Act
-                await wrapper.vm.onFormSubmit();
+                it('should emit `CheckoutValidationError`', async () => {
+                    // Act
+                    await wrapper.vm.onFormSubmit();
 
-                // Assert
-                expect(wrapper.emitted(EventNames.CheckoutValidationError).length).toBe(1);
-                expect(wrapper.emitted(EventNames.CheckoutValidationError)[0][0]).toEqual(mockValidationState);
-            });
-
-            it('should try to call `submitCheckout` if form is Valid', async () => {
-                // Arrange
-                isFormValidSpy.mockReturnValue(true);
-
-                const wrapper = mount(VueCheckout, {
-                    store: createStore(),
-                    i18n,
-                    localVue,
-                    propsData,
-                    mocks: {
-                        $v,
-                        $logger
-                    }
+                    // Assert
+                    expect(wrapper.emitted(EventNames.CheckoutValidationError).length).toBe(1);
                 });
 
-                // Act
-                await wrapper.vm.onFormSubmit();
+                it('should call `trackFormInteraction` with an action `inline_error` and validation errors', async () => {
+                    // Arrange
+                    const payload = {
+                        action: 'inline_error',
+                        error: mockValidationState.invalidFields
+                    };
 
-                // Assert
-                expect(submitCheckoutSpy).toHaveBeenCalled();
+                    // Act
+                    await wrapper.vm.onFormSubmit();
+
+                    // Assert
+                    expect(trackFormInteractionSpy).toHaveBeenCalledWith(payload);
+                });
             });
 
-            it('should call `handleErrorState` if `submitCheckout` returns an error', async () => {
-                // Arrange
-                const handleErrorStateSpy = jest.spyOn(VueCheckout.methods, 'handleErrorState');
+            describe('when the form is valid', () => {
+                let wrapper;
+
+                beforeEach(() => {
+                    isFormValidSpy.mockReturnValue(true);
+
+                    wrapper = mount(VueCheckout, {
+                        store: createStore(),
+                        i18n,
+                        localVue,
+                        propsData,
+                        mocks: {
+                            $v,
+                            $logger
+                        }
+                    });
+                });
+
+                it('should try to call `submitCheckout`', async () => {
+                    // Act
+                    await wrapper.vm.onFormSubmit();
+
+                    // Assert
+                    expect(submitCheckoutSpy).toHaveBeenCalled();
+                });
+
+                it('should call `trackFormInteractionSpy` with correct `action`', async () => {
+                    // Arrange
+                    const payload = {
+                        action: 'success'
+                    };
+
+                    // Act
+                    await wrapper.vm.onFormSubmit();
+
+                    // Assert
+                    expect(trackFormInteractionSpy).toHaveBeenCalledWith(payload);
+                });
+            });
+
+            describe('when `submitCheckout` returns an error', () => {
+                let wrapper;
+                let handleErrorStateSpy;
                 const error = new Error('errorMessage');
 
-                submitCheckoutSpy.mockImplementation(() => { throw error; });
-                isFormValidSpy.mockReturnValue(true);
+                beforeEach(() => {
+                    handleErrorStateSpy = jest.spyOn(VueCheckout.methods, 'handleErrorState');
 
-                const wrapper = mount(VueCheckout, {
-                    store: createStore(),
-                    i18n,
-                    localVue,
-                    propsData,
-                    mocks: {
-                        $v,
-                        $logger
-                    }
+                    submitCheckoutSpy.mockImplementation(() => { throw error; });
+                    isFormValidSpy.mockReturnValue(true);
+
+                    wrapper = mount(VueCheckout, {
+                        store: createStore(),
+                        i18n,
+                        localVue,
+                        propsData,
+                        mocks: {
+                            $v,
+                            $logger
+                        }
+                    });
                 });
 
-                // Act
-                await wrapper.vm.onFormSubmit();
+                it('should call `handleErrorState`', async () => {
+                    // Act
+                    await wrapper.vm.onFormSubmit();
 
-                // Assert
-                expect(handleErrorStateSpy).toHaveBeenCalledWith(error);
+                    // Assert
+                    expect(handleErrorStateSpy).toHaveBeenCalledWith(error);
+                });
             });
 
             it('should call `logWarn` if form is Invalid', async () => {
@@ -2177,7 +2268,7 @@ describe('Checkout', () => {
                 wrapper.vm.isValidPhoneNumber();
 
                 // Assert
-                expect(isValidPhoneNumberSpy).toHaveBeenCalledWith(defaultState.customer.mobileNumber, i18n.locale);
+                expect(isValidPhoneNumberSpy).toHaveBeenCalledWith(defaultCheckoutState.customer.mobileNumber, i18n.locale);
             });
         });
 
@@ -2201,7 +2292,7 @@ describe('Checkout', () => {
                 wrapper.vm.isValidPostcode();
 
                 // Assert
-                expect(isValidPostcodeSpy).toHaveBeenCalledWith(defaultState.address.postcode, i18n.locale);
+                expect(isValidPostcodeSpy).toHaveBeenCalledWith(defaultCheckoutState.address.postcode, i18n.locale);
             });
         });
 
@@ -2252,7 +2343,7 @@ describe('Checkout', () => {
                 jest.advanceTimersByTime(1000);
 
                 // Assert
-                expect(windowLocationSpy).toHaveBeenCalledWith(`${paymentPageUrlPrefix}/${defaultState.orderId}`);
+                expect(windowLocationSpy).toHaveBeenCalledWith(`${paymentPageUrlPrefix}/${defaultCheckoutState.orderId}`);
             });
 
             it('should not redirect to the payment page before 1 second', () => {
@@ -2294,7 +2385,7 @@ describe('Checkout', () => {
                         basketId,
                         applicationId: 7,
                         customerNotes: {
-                            noteForRestaurant: defaultState.userNote
+                            noteForRestaurant: defaultCheckoutState.userNote
                         },
                         applicationName,
                         applicationVersion: '1',
@@ -2307,7 +2398,7 @@ describe('Checkout', () => {
 
                 const wrapper = shallowMount(VueCheckout, {
                     store: createStore({
-                        ...defaultState,
+                        ...defaultCheckoutState,
                         basket: {
                             id: basketId
                         }