+++ packages/components/molecules/f-alert/CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 Latest (add to next release)
 ------------------------------
+*February 25, 2021*
+
+### Changed
+- Restructured component object into page object model
+- Refactored component and accessibility tests
+
+
 *January 12, 2021*
 
 ### Changed