+++ packages/components/organisms/f-checkout/CHANGELOG.md
@@ -15,6 +15,17 @@ v0.60.0
 - Mapped the results into the `PATCH` Checkout request model.
 - Refactored/Added unit tests.
 
+v0.60.0
+-------------------------------
+*February 26, 2021*
+
+### Added
+- `analytics.module` to store.
+- `mapAnalyticsName` and `mapAnalyticsNames` to `mapper` service.
+- Analytics tracking to `Checkout` component.
+- Tests to cover changes.
+
+
 v0.59.0
 ------------------------------
 *February 24, 2021*